Create JavascriptSpeedup
Some of this can be removed because I wrote it years ago but basically what this script does it loads all the pages on landing so that the user does see a delay due to reload when switching pages or tabs and grabs them by their headers instead of the page location. in short everything is loaded and then the JS just switches out the main page for others.